16/6

+++ Lucy_SalesData.Models/OrderDetail.cs
@@ -6,6 +6,6 @@
         public int ProductID { get; set; }
         public double UnitPrice { get; set; }
         public int Quantity { get; set; }
-        public float Discount { get; set; }
+        public double Discount { get; set; }
     }
 }
